requirements.txt
cycler==0.11.0
fonttools==4.29.1
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
networkx==2.6.3
numpy==1.22.2
packaging==21.3
pandas==1.4.1
patsy==0.5.2
Pillow==9.0.1
pyfuzzylite==7.0b3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
scikit-fuzzy==0.4.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
threadpoolctl==3.1.0
